# api/storage/openapi.yaml
openapi: 3.0.0
info:
  title: 'Mordomus API'
  description: 'API do Mordomus — gestão doméstica. Todo o conteúdo vive sob `/api/v1`.'
  version: 1.0.0
servers:
  -
    url: /
    description: 'Origem do gateway (nginx) que faz proxy para o monólito'
paths:
...
  /api/v1/scheduling/occurrences:
    get:
      tags:
        - scheduling
      summary: 'Lista as ocorrências da residência ativa'
      description: 'Ordena por `due_at`. `from`/`to` são dias de calendário no fuso da residência e `to` é inclusivo; sem eles a agenda inteira é devolvida, paginada.'
      operationId: 34bfae6fd3a2c04e0baf2eeaf72a56ad
      parameters:
        -
          parameter: page
          name: index
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
        -
          parameter: per_page
          name: index
          in: query
          schema:
            type: integer
            default: 25
            maximum: 100
            minimum: 1
        -
          parameter: from
          name: index
          in: query
          schema:
            type: string
            format: date
            example: '2026-10-01'
        -
          parameter: to
          name: index
          in: query
          schema:
            type: string
            format: date
            example: '2026-10-31'
        -
          parameter: subject_type
          name: index
          in: query
          description: 'Filtra por tipo de alvo'
          schema:
            type: string
            enum:
              - asset
              - bill
        -
          parameter: status
          name: index
          in: query
          schema:
            type: string
            enum:
              - pending
              - notified
              - completed
              - skipped
              - overdue
      responses:
        '200':
          description: 'Ocorrências paginadas'
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Occurrence' } }
                  meta: { $ref: '#/components/schemas/PageMeta' }
                type: object
        '401':
          description: 'Token ausente ou inválido'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'Sem residência ativa no token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 'Filtro inválido'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 'Rate limit do gateway'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 'Erro interno'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        -
          jwtBearerAuth: []
  '/api/v1/scheduling/occurrences/{occurrence}/complete':
    post:
      tags:
        - scheduling
      summary: 'Conclui uma ocorrência e recalcula o ciclo'
      description: 'Idempotente: concluir de novo devolve a mesma ocorrência sem abrir um segundo ciclo. O recálculo roda na fila e publica `occurrence.completed` com a próxima data.'
      operationId: 90ac2195920cced5e5ada4cc7ff41066
      parameters:
        -
          parameter: occurrence
          name: complete
          in: path
          description: 'Id ULID da ocorrência'
          required: true
          schema:
            type: string
            maxLength: 26
      responses:
        '200':
          description: 'Ocorrência concluída; a próxima data vem no evento `occurrence.completed`'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Occurrence' }
                type: object
        '401':
          description: 'Token ausente ou inválido'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'Capability occurrences.complete ausente'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'Ocorrência não encontrada na residência ativa'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 'Ocorrência já dispensada'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 'Dados inválidos'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 'Rate limit do gateway'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 'Erro interno'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        -
          jwtBearerAuth: []
  '/api/v1/scheduling/occurrences/{occurrence}/skip':
    post:
      tags:
        - scheduling
      summary: 'Dispensa uma ocorrência sem concluir o ciclo'
      description: 'A linha vira `skipped` e permanece como histórico. O ciclo não é recalculado: pular diz que aquele dia não acontece, e a próxima data continua a que a regra já tinha.'
      operationId: 5a6382fb32b391b2e0de26d73930844d
      parameters:
        -
          parameter: occurrence
          name: skip
          in: path
          description: 'Id ULID da ocorrência'
          required: true
          schema:
            type: string
            maxLength: 26
      responses:
        '200':
          description: 'Ocorrência dispensada'
          content:
            application/json:
              schema:
                properties:
                  data: { $ref: '#/components/schemas/Occurrence' }
                type: object
        '401':
          description: 'Token ausente ou inválido'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 'Capability occurrences.skip ausente'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 'Ocorrência não encontrada na residência ativa'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: 'Ocorrência já concluída'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: 'Dados inválidos'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: 'Rate limit do gateway'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 'Erro interno'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        -
          jwtBearerAuth: []
...
components:
  schemas:
...
    Error:
      description: 'Envelope de erro da API'
      required:
        - error
      properties:
        error:
          required:
            - code
            - message
            - details
            - request_id
          properties:
            code:
              type: string
              example: validation_failed
            message:
              type: string
              example: 'Dados inválidos.'
            details:
              description: 'Detalhes livres: objeto por campo (validação) ou lista de esperados/recebidos'
              oneOf:
                -
                  type: object
                  example: { email: ['O e-mail já está em uso.'] }
                  additionalProperties: true
                -
                  type: array
                  items: { type: string }
                  example: [01ABC]
            request_id:
              type: string
              example: smoke-0001
              nullable: true
          type: object
      type: object
...
    Occurrence:
      description: 'Ocorrência materializada de uma regra: dia do calendário, instante de disparo e situação no ciclo'
      required:
        - id
        - tenant_id
        - trigger_config_id
        - subject_type
        - subject_id
        - title
        - scheduled_for
        - due_at
        - status
      properties:
        id:
          type: string
          example: 01J8Z0M9W3K6Q2T4R5Y7B8C9D2
        tenant_id:
          type: string
          example: 01J8Z0M9W3K6Q2T4R5Y7B8C9D1
        trigger_config_id:
          description: 'Regra que gerou a ocorrência'
          type: string
        subject_type:
          description: 'Alvo da regra; sempre presente — a FK leva a ocorrência junto na cascata'
          type: string
          enum:
            - asset
            - bill
        subject_id:
          description: 'Ativo ou conta alvo, conforme subject_type'
          type: string
        title:
          type: string
          example: 'Trocar o filtro do ar'
        scheduled_for:
          description: 'Dia do calendário a que a ocorrência pertence'
          type: string
          format: date
          example: '2026-10-15'
        due_at:
          description: 'Instante de disparo, na hora preferida da residência'
          type: string
          format: date-time
          example: '2026-10-15T12:00:00Z'
        status:
          description: 'overdue é constatação de atraso e ainda aceita check-in'
          type: string
          enum:
            - pending
            - notified
            - completed
            - skipped
            - overdue
        notified_at:
          description: 'Primeiro aviso publicado para esta ocorrência'
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        completed_by:
          description: 'Morador que fechou o ciclo; nulo para ocorrência gerada pela fila'
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
      type: object
    PageMeta:
      description: 'Paginação offset'
      required:
        - page
        - per_page
        - total
        - last_page
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 25
        total:
          type: integer
          example: 42
        last_page:
          type: integer
          example: 2
      type: object
...
  securitySchemes:
    jwtBearerAuth:
      type: http
      description: 'Access token devolvido por `/identity/auth/login` e demais rotas de sessão.'
      bearerFormat: JWT
      scheme: bearer
...